refactor(paywalls): declare paywallPresentationID next to the view publishing it
One `UUID?` read by one component did not warrant its own file. Moves the key
alongside `SafeAreaInsetsKey` at the bottom of `PaywallsV2View`, which already
follows this pattern and is inside the same `!os(tvOS)` guard.
